@@ -1,19 +1,19 @@
 {
   "name": "@candiman/website",
-  "version": "0.0.7-beta.12",
+  "version": "0.0.8",
   "schematics": "./schematics/collection.json",
   "ng-update": {
     "migrations": "./schematics/migration.json"
   },
   "peerDependencies": {
-    "@angular/common": "^7.1.0",
-    "@angular/core": "^7.1.0",
-    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
-    "@fortawesome/angular-fontawesome": "0.1.0-9",
-    "@fortawesome/fontawesome-svg-core": "^1.2.0-11",
-    "@fortawesome/free-solid-svg-icons": "^5.1.0-8",
-    "@fortawesome/free-regular-svg-icons": "^5.1.0-8",
-    "@fortawesome/free-brands-svg-icons": "5.1.0-8",
-    "bootstrap": "^4.0.0"
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@fortawesome/angular-fontawesome": "0.4.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.18",
+    "@fortawesome/free-solid-svg-icons": "^5.8.2",
+    "@fortawesome/free-regular-svg-icons": "^5.8.2",
+    "@fortawesome/free-brands-svg-icons": "5.8.2",
+    "bootstrap": "^4.3.1"
   }
 }
